requirements.txt
asgiref==3.8.1
Django==5.1.2
django-appconf==1.0.6
django-ckeditor-5==0.2.15
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
